Made general Animations class shared across entities, and a specific AnimationComponent for player.

+++ Src/main.cpp
@@ -15,7 +15,7 @@ int main() {
 
     #pragma region world
     const auto worldForest = game.createWorld("Forest");
-    worldForest->groundLevel = window.getSize().y*0.9f/2.f;
+    worldForest->groundLevel = static_cast<float>(window.getSize().y)*0.9f/2.f;
     #pragma endregion
 
     #pragma region background
